# [xy] Cache project platform settings data in API calls.