--- a/mage_ai/api/resources/PipelineResource.py
+++ b/mage_ai/api/resources/PipelineResource.py
@@ -427,6 +427,8 @@ class PipelineResource(BaseResource):
     @classmethod
     @safe_db_query
     async def create(self, payload, user, **kwargs):
+        context_data = kwargs.get('context_data')
+
         clone_pipeline_uuid = payload.get('clone_pipeline_uuid')
         template_uuid = payload.get('custom_template_uuid')
         name = payload.get('name')
@@ -436,7 +438,7 @@ class PipelineResource(BaseResource):
         llm_payload = payload.get('llm')
         pipeline = None
 
-        repo_path = get_repo_path(user=user)
+        repo_path = get_repo_path(context_data=context_data, user=user)
         if template_uuid:
             custom_template = CustomPipelineTemplate.load(
                 repo_path,
@@ -526,7 +528,10 @@ class PipelineResource(BaseResource):
                             block.update(dict(upstream_blocks=arr))
 
         if pipeline:
-            await UsageStatisticLogger().pipeline_create(
+            await UsageStatisticLogger(
+                context_data=context_data,
+                repo_path=repo_path,
+            ).pipeline_create(
                 pipeline,
                 clone_pipeline_uuid=clone_pipeline_uuid,
                 llm_payload=llm_payload,
@@ -553,6 +558,7 @@ class PipelineResource(BaseResource):
 
         self.on_create_callback = _on_create_callback
 
+        pipeline.context_data = context_data
         return self(pipeline, user, **kwargs)
 
     @classmethod
@@ -564,6 +570,7 @@ class PipelineResource(BaseResource):
             return await get_pipeline_from_platform_async(
                 pipeline_uuid,
                 repo_path=repo_path,
+                context_data=kwargs.get('context_data'),
             )
 
         return await Pipeline.get_async(
@@ -587,7 +594,9 @@ class PipelineResource(BaseResource):
     @classmethod
     @safe_db_query
     async def member(self, pk, user, **kwargs):
-        repo_path = get_repo_path(user=user)
+        context_data = kwargs.get('context_data')
+
+        repo_path = get_repo_path(context_data=context_data, user=user)
         pipeline = await self.__fetch_model(pk, repo_path, **kwargs)
 
         api_operation_action = kwargs.get('api_operation_action', None)
@@ -601,7 +610,10 @@ class PipelineResource(BaseResource):
             )
 
         if api_operation_action == DETAIL:
-            if Project(repo_config=pipeline.repo_config).is_feature_enabled(
+            if Project(
+                context_data=context_data,
+                repo_config=pipeline.repo_config,
+            ).is_feature_enabled(
                 FeatureUUID.OPERATION_HISTORY,
             ):
                 record_detail_pipeline(
@@ -628,6 +640,7 @@ class PipelineResource(BaseResource):
             if mapping.get(pipeline.uuid):
                 pipeline.schedules = mapping[pipeline.uuid] or []
 
+        pipeline.context_data = context_data
         return self(pipeline, user, **kwargs)
 
     @safe_db_query
@@ -663,6 +676,7 @@ class PipelineResource(BaseResource):
 
     @safe_db_query
     async def update(self, payload, **kwargs):
+        context_data = kwargs.get('context_data')
         if 'add_upstream_for_block_uuid' in payload:
             block_uuid = payload['add_upstream_for_block_uuid']
             block = self.model.get_block(block_uuid, widget=False)
@@ -675,7 +689,7 @@ class PipelineResource(BaseResource):
                 self.model.save()
             return self
 
-        repo_path = get_repo_path(user=self.current_user)
+        repo_path = get_repo_path(context_data=context_data, user=self.current_user)
 
         query = kwargs.get('query', {})
         update_content = query.get('update_content', [False])