--- a/mage_ai/api/resources/VariableResource.py
+++ b/mage_ai/api/resources/VariableResource.py
@@ -10,7 +10,7 @@ from mage_ai.data_preparation.variable_manager import (
     set_global_variable,
 )
 from mage_ai.orchestration.db import safe_db_query
-from mage_ai.settings.repo import get_variables_dir
+from mage_ai.settings.repo import get_repo_path, get_variables_dir
 
 
 def get_variable_value(
@@ -25,8 +25,8 @@ def get_variable_value(
         variable_uuid,
     )
     if variable.variable_type in [VariableType.DATAFRAME, VariableType.GEO_DATAFRAME]:
-        value = "DataFrame"
-        variable_type = "pandas.DataFrame"
+        value = 'DataFrame'
+        variable_type = 'pandas.DataFrame'
     else:
         value = variable.read_data(sample=True)
         variable_type = str(type(value))
@@ -41,27 +41,38 @@ class VariableResource(GenericResource):
     @classmethod
     @safe_db_query
     def collection(self, query, meta, user, **kwargs):
-        pipeline_uuid = kwargs["parent_model"].uuid
+        context_data = kwargs.get('context_data')
+        repo_path = get_repo_path(context_data=context_data, user=user)
+        variables_dir = get_variables_dir(repo_path=repo_path)
+        pipeline_uuid = kwargs['parent_model'].uuid
 
-        global_only = query.get("global_only", [False])
+        global_only = query.get('global_only', [False])
         if global_only:
             global_only = global_only[0]
 
         # Get global variables from project's path
         global_variables = [
             dict(uuid=uuid, type=str(type(value)), value=value)
-            for uuid, value in get_global_variables(pipeline_uuid).items()
+            for uuid, value in get_global_variables(
+                pipeline_uuid,
+                context_data=context_data,
+                repo_path=repo_path,
+                variables_dir=variables_dir,
+            ).items()
         ]
         global_variables_arr = [
             dict(
-                block=dict(uuid="global"),
+                block=dict(uuid='global'),
                 pipeline=dict(uuid=pipeline_uuid),
                 variables=global_variables,
             )
         ]
         variables = global_variables_arr
         if not global_only:
-            variable_manager = VariableManager(variables_dir=get_variables_dir())
+            variable_manager = VariableManager(
+                repo_path=repo_path,
+                variables_dir=variables_dir,
+            )
             variables_dict = variable_manager.get_variables_by_pipeline(pipeline_uuid)
             variables = [
                 dict(
@@ -76,11 +87,11 @@ class VariableResource(GenericResource):
                         )
                         for var in arr
                         # Not return printed outputs
-                        if var == "df" or var.startswith("output")
+                        if var == 'df' or var.startswith('output')
                     ],
                 )
                 for uuid, arr in variables_dict.items()
-                if uuid != "global"
+                if uuid != 'global'
             ] + global_variables_arr
 
         return self.build_result_set(
@@ -91,21 +102,21 @@ class VariableResource(GenericResource):
 
     @classmethod
     @safe_db_query
-    def create(self, payload: Dict, user, **kwargs) -> "VariableResource":
-        pipeline_uuid = kwargs["parent_model"].uuid
+    def create(self, payload: Dict, user, **kwargs) -> 'VariableResource':
+        pipeline_uuid = kwargs['parent_model'].uuid
 
         error = ApiError.RESOURCE_INVALID.copy()
 
-        variable_uuid = payload.get("name")
+        variable_uuid = payload.get('name')
         if not variable_uuid.isidentifier():
             error.update(
-                message=f"Invalid variable name syntax for variable name {variable_uuid}."
+                message=f'Invalid variable name syntax for variable name {variable_uuid}.'
             )
             raise ApiError(error)
 
-        variable_value = payload.get("value")
+        variable_value = payload.get('value')
         if variable_value is None:
-            error.update(message=f"Value is empty for variable name {variable_uuid}.")
+            error.update(message=f'Value is empty for variable name {variable_uuid}.')
             raise ApiError(error)
 
         set_global_variable(
@@ -118,7 +129,7 @@ class VariableResource(GenericResource):
 
         return self(
             dict(
-                block=dict(uuid="global"),
+                block=dict(uuid='global'),
                 name=variable_uuid,
                 pipeline=dict(uuid=pipeline_uuid),
                 value=variable_value,
@@ -135,20 +146,20 @@ class VariableResource(GenericResource):
 
     @safe_db_query
     def update(self, payload, **kwargs):
-        pipeline_uuid = kwargs["parent_model"].uuid
+        pipeline_uuid = kwargs['parent_model'].uuid
 
         error = ApiError.RESOURCE_INVALID.copy()
 
-        variable_uuid = payload.get("name")
+        variable_uuid = payload.get('name')
         if not variable_uuid.isidentifier():
             error.update(
-                message=f"Invalid variable name syntax for variable name {self.name}."
+                message=f'Invalid variable name syntax for variable name {self.name}.'
             )
             raise ApiError(error)
 
-        variable_value = payload.get("value")
+        variable_value = payload.get('value')
         if variable_value is None:
-            error.update(message=f"Value is empty for variable name {self.name}.")
+            error.update(message=f'Value is empty for variable name {self.name}.')
             raise ApiError(error)
 
         set_global_variable(
@@ -161,7 +172,7 @@ class VariableResource(GenericResource):
 
         self.model.update(
             dict(
-                block=dict(uuid="global"),
+                block=dict(uuid='global'),
                 name=variable_uuid,
                 pipeline=dict(uuid=pipeline_uuid),
                 value=variable_value,
@@ -173,6 +184,6 @@ class VariableResource(GenericResource):
 
     @safe_db_query
     def delete(self, **kwargs):
-        pipeline_uuid = kwargs["parent_model"].uuid
+        pipeline_uuid = kwargs['parent_model'].uuid
         delete_global_variable(pipeline_uuid, self.name)
         return self