# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
Cache project platform settings data in API calls to speed up API calls


# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] pipeline LIST: reduce time from 270ms to 90ms
- [x] status: 100ms -> 62ms 
- [x] Pipeline GET: 280ms -> 100ms 
- [x] Pipeline schedule LIST:  118ms -> 70ms 
- [x] Pipeline trigger LIST: 130ms -> 70ms 
- [x] Variable LIST: 280ms -> 80ms 
- [x] GitBranch GET: 170ms -> 60ms 
- [x] Project GET: 380ms -> 300ms 


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature, documentation)
- [x] I have performed a self-review of my own code
- [x] I have added unit tests that prove my fix is effective or that my feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request -->